feat: 정수 외에 소수도 계산 가능하도록 수정

+++ src/main/java/caculator/controller/CalculatorController.java
@@ -13,7 +13,7 @@ public class CalculatorController {
         PostfixConvertor postfixConvertor = new PostfixConvertor();
         List<String> postfixFormula = postfixConvertor.convertPostfixFormula(formula);
         Calculator calculator = new Calculator(postfixFormula);
-        Integer result = calculator.calculate();
+        Double result = calculator.calculate();
         OutputView.printCalculatorResult(result);
     }
 }
