+++ src/main/java/caculator/views/OutputView.java
@@ -2,7 +2,7 @@ package caculator.views;
 
 public class OutputView {
 
-    public static void printCalculatorResult(Integer result) {
-        System.out.println(result.intValue());
+    public static void printCalculatorResult(Double result) {
+        System.out.println(result.doubleValue());
     }
 }
